+++ docs/scripts/views/button.vue
@@ -118,7 +118,8 @@
     </div>
     <ui-code language="html" :code="code9.html"></ui-code>
 
-    <ui-tabs effect :active="tab" @switched="onChange">
+    <h4>Button API</h4>
+    <ui-tabs effect position="left" :active="tab" @switched="onChange">
       <ui-panel tab="props">
         <ui-table
           class="docs-table"